requirements.txt
alembic==1.7.6
aniso8601==9.0.1
click==8.1.7
Flask==2.0.3
Flask-RESTful==0.3.10
Flask-SQLAlchemy==2.5.1
greenlet==3.0.1
importlib-metadata==6.8.0
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.3.0
MarkupSafe==2.1.3
psycopg2-binary==2.9.3
PyJWT==2.3.0
python-dotenv==0.19.2
pytz==2023.3.post1
six==1.16.0
SQLAlchemy==1.4.31
Werkzeug==2.2.2
zipp==3.17.0
